requirements.txt
flask==2.0.1
instagrapi==1.16.20
python-dotenv==1.0.0
openai==1.3.0
Pillow==9.5.0
python-telegram-bot==20.6 